# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dashboard
  - name: tickets
  - name: customers
  - name: public
  - name: stores
  - name: auth
  - name: admin
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get operational dashboard
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
  /stores:
    get:
      operationId: listStores
      tags: [stores]
      summary: List stores available to the signed-in operator
      responses:
        "200":
          description: Store list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Store" }
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginInput" }
      responses:
        "200":
          description: Signed-in user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionUser" }
        "401": { description: Invalid credentials }
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Sign out and clear session cookie
      responses:
        "204": { description: Signed out }
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Current signed-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionUser" }
        "401": { description: Not signed in }
  /admin/stores:
    get:
      operationId: listAdminStores
      tags: [admin]
      summary: List all stores and membership status
      responses:
        "200":
          description: Admin store list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminStore" }
        "403": { description: Admin only }
    post:
      operationId: createAdminStore
      tags: [admin]
      summary: Create a store, membership, and first operator
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateStoreInput" }
      responses:
        "201":
          description: Created store
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminStore" }
        "409": { description: Store code or operator email already exists }
  /admin/stores/{id}/membership:
    patch:
      operationId: updateStoreMembership
      tags: [admin]
      summary: Update store membership plan and payment status
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateMembershipInput" }
      responses:
        "200":
          description: Updated store
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminStore" }
        "404": { description: Store not found }
  /admin/stores/{id}/operators:
    post:
      operationId: createStoreOperator
      tags: [admin]
      summary: Add an operator login for a store
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOperatorInput" }
      responses:
        "201":
          description: Created operator
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StoreOperator" }
        "404": { description: Store not found }
        "409": { description: Email already exists }
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List and filter repair tickets
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
        - { name: technicianId, in: query, schema: { type: string } }
        - { name: priority, in: query, schema: { type: string } }
      responses:
        "200":
          description: Ticket list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Ticket" }
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create a repair ticket
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TicketInput" }
      responses:
        "201":
          description: Created ticket
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get ticket detail
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Ticket detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TicketDetail" }
        "404": { description: Ticket not found }
    patch:
      operationId: updateTicket
      tags: [tickets]
      summary: Update ticket
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TicketUpdate" }
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TicketDetail" }
  /tickets/{id}/status:
    post:
      operationId: updateTicketStatus
      tags: [tickets]
      summary: Change status and append audit history
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StatusUpdate" }
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TicketDetail" }
  /tickets/{id}/payments:
    post:
      operationId: recordPayment
      tags: [tickets]
      summary: Record a payment
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PaymentInput" }
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Payment" }
  /tickets/{id}/notifications/{notificationId}/retry:
    post:
      operationId: retryNotification
      tags: [tickets]
      summary: Retry a failed WhatsApp notification
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: notificationId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Retried notification
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Notification" }
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      responses:
        "200":
          description: Customer list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Customer" }
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CustomerInput" }
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
  /public/track/{ticketNumber}:
    get:
      operationId: getPublicTracking
      tags: [public]
      summary: Public ticket tracking
      parameters:
        - { name: ticketNumber, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Public tracking data
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicTracking" }
        "404": { description: Ticket not found }
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Dashboard:
      type: object
      required: [totalToday, newTickets, underRepair, waitingParts, delayed, completed, readyPickup, recentTickets, upcomingDeadlines, statusCounts]
      properties:
        totalToday: { type: number }
        newTickets: { type: number }
        underRepair: { type: number }
        waitingParts: { type: number }
        delayed: { type: number }
        completed: { type: number }
        readyPickup: { type: number }
        recentTickets: { type: array, items: { $ref: "#/components/schemas/Ticket" } }
        upcomingDeadlines: { type: array, items: { $ref: "#/components/schemas/Ticket" } }
        statusCounts: { type: object, additionalProperties: { type: number } }
    Store:
      type: object
      required: [id, name, code, address, phone, active, membershipPlan, membershipExpiresAt]
      properties:
        id: { type: string }
        name: { type: string }
        code: { type: string }
        address: { type: string }
        phone: { type: string }
        active: { type: boolean }
        membershipPlan: { type: string, enum: [unpaid, monthly, lifetime] }
        membershipExpiresAt: { type: ["string", "null"] }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    SessionUser:
      type: object
      required: [id, name, email, role, stores]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [platform_admin, store_operator] }
        stores:
          type: array
          items: { $ref: "#/components/schemas/Store" }
    StoreOperator:
      type: object
      required: [id, name, email]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
    AdminStore:
      allOf:
        - $ref: "#/components/schemas/Store"
        - type: object
          required: [operators]
          properties:
            operators:
              type: array
              items: { $ref: "#/components/schemas/StoreOperator" }
    CreateStoreInput:
      type: object
      required: [name, code, address, phone, membershipPlan, operatorName, operatorEmail, operatorPassword]
      properties:
        name: { type: string }
        code: { type: string }
        address: { type: string }
        phone: { type: string }
        membershipPlan: { type: string, enum: [unpaid, monthly, lifetime] }
        membershipExpiresAt: { type: string }
        operatorName: { type: string }
        operatorEmail: { type: string }
        operatorPassword: { type: string }
    UpdateMembershipInput:
      type: object
      required: [membershipPlan]
      properties:
        membershipPlan: { type: string, enum: [unpaid, monthly, lifetime] }
        membershipExpiresAt: { type: string }
    CreateOperatorInput:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }
    Customer:
      type: object
      required: [id, name, whatsapp, totalTickets, activeTickets]
      properties:
        id: { type: string }
        name: { type: string }
        whatsapp: { type: string }
        email: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        totalTickets: { type: number }
        activeTickets: { type: number }
    CustomerInput:
      type: object
      required: [name, whatsapp]
      properties:
        name: { type: string }
        whatsapp: { type: string }
        email: { type: string }
        address: { type: string }
    Ticket:
      type: object
      required: [id, ticketNumber, customerName, whatsapp, deviceBrand, deviceModel, complaint, status, priority, intakeDate, estimatedCompletion, technicianName, totalCost, paidAmount]
      properties:
        id: { type: string }
        ticketNumber: { type: string }
        customerName: { type: string }
        whatsapp: { type: string }
        deviceBrand: { type: string }
        deviceModel: { type: string }
        imei: { type: ["string", "null"] }
        color: { type: ["string", "null"] }
        complaint: { type: string }
        status: { type: string }
        priority: { type: string }
        intakeDate: { type: string }
        estimatedCompletion: { type: ["string", "null"] }
        technicianName: { type: ["string", "null"] }
        totalCost: { type: number }
        paidAmount: { type: number }
        unreadNotifications: { type: number }
    TicketInput:
      type: object
      required: [customerName, whatsapp, deviceBrand, deviceModel, complaint, priority]
      properties:
        customerId: { type: string }
        customerName: { type: string }
        whatsapp: { type: string }
        deviceBrand: { type: string }
        deviceModel: { type: string }
        imei: { type: string }
        color: { type: string }
        storage: { type: string }
        condition: { type: string }
        accessories: { type: array, items: { type: string } }
        complaint: { type: string }
        priority: { type: string }
        estimatedCompletion: { type: string }
        technicianId: { type: string }
        notes: { type: string }
    TicketUpdate:
      type: object
      properties:
        technicianId: { type: string }
        priority: { type: string }
        estimatedCompletion: { type: string }
        diagnosis: { type: string }
        cause: { type: string }
        recommendation: { type: string }
        laborCost: { type: number }
        partsCost: { type: number }
        delayReason: { type: string }
        notes: { type: string }
    StatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string }
        note: { type: string }
        delayReason: { type: string }
    Payment:
      type: object
      required: [id, amount, method, paidAt]
      properties:
        id: { type: string }
        amount: { type: number }
        method: { type: string }
        paidAt: { type: string }
        receivedBy: { type: string }
    PaymentInput:
      type: object
      required: [amount, method]
      properties:
        amount: { type: number }
        method: { type: string }
    Notification:
      type: object
      required: [id, type, recipient, status, sentAt, message]
      properties:
        id: { type: string }
        type: { type: string }
        recipient: { type: string }
        status: { type: string }
        sentAt: { type: ["string", "null"] }
        message: { type: string }
        error: { type: ["string", "null"] }
    StatusHistory:
      type: object
      required: [id, status, at, actor]
      properties:
        id: { type: string }
        status: { type: string }
        at: { type: string }
        actor: { type: string }
        note: { type: ["string", "null"] }
    TicketDetail:
      allOf:
        - $ref: "#/components/schemas/Ticket"
        - type: object
          required: [condition, accessories, diagnosis, cause, recommendation, laborCost, partsCost, payments, notifications, history, outstandingBalance]
          properties:
            condition: { type: ["string", "null"] }
            accessories: { type: array, items: { type: string } }
            diagnosis: { type: ["string", "null"] }
            cause: { type: ["string", "null"] }
            recommendation: { type: ["string", "null"] }
            laborCost: { type: number }
            partsCost: { type: number }
            payments: { type: array, items: { $ref: "#/components/schemas/Payment" } }
            notifications: { type: array, items: { $ref: "#/components/schemas/Notification" } }
            history: { type: array, items: { $ref: "#/components/schemas/StatusHistory" } }
            outstandingBalance: { type: number }
    PublicTracking:
      type: object
      required: [ticketNumber, customerName, deviceBrand, deviceModel, complaint, status, history, estimatedCompletion]
      properties:
        ticketNumber: { type: string }
        customerName: { type: string }
        deviceBrand: { type: string }
        deviceModel: { type: string }
        complaint: { type: string }
        status: { type: string }
        history: { type: array, items: { $ref: "#/components/schemas/StatusHistory" } }
        estimatedCompletion: { type: ["string", "null"] }
        delayReason: { type: ["string", "null"] }
        pickupInfo: { type: ["string", "null"] }