# lib/api-zod/src/generated/types/adminStore.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Store } from './store';
import type { StoreOperator } from './storeOperator';

export type AdminStore = Store & {
  operators: StoreOperator[];
};