# lib/api-zod/src/generated/types/createOperatorInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOperatorInput {
  name: string;
  email: string;
  password: string;
}