# lib/api-zod/src/generated/types/createStoreInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateStoreInputMembershipPlan } from './createStoreInputMembershipPlan';

export interface CreateStoreInput {
  name: string;
  code: string;
  address: string;
  phone: string;
  membershipPlan: CreateStoreInputMembershipPlan;
  membershipExpiresAt?: string;
  operatorName: string;
  operatorEmail: string;
  operatorPassword: string;
}