# lib/api-zod/src/generated/types/sessionUser.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SessionUserRole } from './sessionUserRole';
import type { Store } from './store';

export interface SessionUser {
  id: string;
  name: string;
  email: string;
  role: SessionUserRole;
  stores: Store[];
}