# lib/api-zod/src/generated/types/store.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StoreMembershipPlan } from './storeMembershipPlan';

export interface Store {
  id: string;
  name: string;
  code: string;
  address: string;
  phone: string;
  active: boolean;
  membershipPlan: StoreMembershipPlan;
  /** @nullable */
  membershipExpiresAt: string | null;
}