# lib/api-zod/src/generated/types/updateMembershipInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateMembershipInputMembershipPlan } from './updateMembershipInputMembershipPlan';

export interface UpdateMembershipInput {
  membershipPlan: UpdateMembershipInputMembershipPlan;
  membershipExpiresAt?: string;
}